+++ env.d.ts
@@ -21,7 +21,7 @@ declare module "obsidian" {
   }
 
   interface Menu {
-    readonly items: { title: string; icon: string; section: string; warning: boolean }[];
+    readonly items: { title: string; icon: string; checked: boolean | null; section: string; warning: boolean }[];
     showAtMouseEventCalls: MouseEvent[];
     pick(index: number): Promise<void>;
   }